feat: Update build process and documentation for pyhl on Linux and Windows, functioning Windows pyhl!
